Master search bar and misc GUI clean-up
