+++ GUI/Localization/zh-CN/Main.zh-CN.resx
@@ -183,51 +183,7 @@
   <data name="reportClientIssueToolStripMenuItem.Text" xml:space="preserve"><value>报告CKAN客户端的问题</value></data>
   <data name="reportMetadataIssueToolStripMenuItem.Text" xml:space="preserve"><value>报告Mod元数据的问题</value></data>
   <data name="aboutToolStripMenuItem.Text" xml:space="preserve"><value>关于</value></data>
-  <data name="launchKSPToolStripMenuItem.Text" xml:space="preserve"><value>启动KSP</value></data>
-  <data name="RefreshToolButton.Text" xml:space="preserve"><value>刷新</value></data>
-  <data name="UpdateAllToolButton.Text" xml:space="preserve"><value>添加可用更新</value></data>
-  <data name="ApplyToolButton.Text" xml:space="preserve"><value>应用更改</value></data>
-  <data name="FilterToolButton.Text" xml:space="preserve"><value>筛选 (兼容)</value></data>
-  <data name="FilterCompatibleButton.Text" xml:space="preserve"><value>兼容</value></data>
-  <data name="FilterInstalledButton.Text" xml:space="preserve"><value>已安装</value></data>
-  <data name="FilterInstalledUpdateButton.Text" xml:space="preserve"><value>已安装 (更新可用)</value></data>
-  <data name="FilterReplaceableButton.Text" xml:space="preserve"><value>可替换</value></data>
-  <data name="FilterCachedButton.Text" xml:space="preserve"><value>已缓存</value></data>
-  <data name="FilterUncachedButton.Text" xml:space="preserve"><value>未缓存</value></data>
-  <data name="FilterNewButton.Text" xml:space="preserve"><value>新兼容</value></data>
-  <data name="FilterNotInstalledButton.Text" xml:space="preserve"><value>未安装</value></data>
-  <data name="FilterIncompatibleButton.Text" xml:space="preserve"><value>不兼容</value></data>
-  <data name="FilterAllButton.Text" xml:space="preserve"><value>所有</value></data>
-  <data name="FilterTagsToolButton.Text" xml:space="preserve"><value>分类</value></data>
-  <data name="FilterLabelsToolButton.Text" xml:space="preserve"><value>标签</value></data>
-  <data name="NavBackwardToolButton.ToolTipText" xml:space="preserve"><value>上一个选中Mod...</value></data>
-  <data name="NavForwardToolButton.ToolTipText" xml:space="preserve"><value>下一个选中Mod...</value></data>
-  <data name="Installed.HeaderText" xml:space="preserve"><value>    安装</value></data>
-  <data name="AutoInstalled.HeaderText" xml:space="preserve"><value>自动安装</value></data>
-  <data name="UpdateCol.HeaderText" xml:space="preserve"><value>更新</value></data>
-  <data name="ReplaceCol.HeaderText" xml:space="preserve"><value>替换</value></data>
-  <data name="ModName.HeaderText" xml:space="preserve"><value>名称</value></data>
-  <data name="Author.HeaderText" xml:space="preserve"><value>作者</value></data>
-  <data name="InstalledVersion.HeaderText" xml:space="preserve"><value>已安装版本</value></data>
-  <data name="LatestVersion.HeaderText" xml:space="preserve"><value>最新版本</value></data>
-  <data name="KSPCompatibility.HeaderText" xml:space="preserve"><value>最高KSP版本</value></data>
-  <data name="SizeCol.HeaderText" xml:space="preserve"><value>下载大小</value></data>
-  <data name="InstallDate.HeaderText" xml:space="preserve"><value>安装日期</value></data>
-  <data name="DownloadCount.HeaderText" xml:space="preserve"><value>下载量</value></data>
-  <data name="Description.HeaderText" xml:space="preserve"><value>描述</value></data>
-  <data name="reinstallToolStripMenuItem.Text" xml:space="preserve"><value>重新安装</value></data>
-  <data name="downloadContentsToolStripMenuItem.Text" xml:space="preserve"><value>下载内容</value></data>
-  <data name="purgeContentsToolStripMenuItem.Text" xml:space="preserve"><value>清除内容</value></data>
   <data name="ManageModsTabPage.Text" xml:space="preserve"><value>管理模组</value></data>
-  <data name="FilterByAuthorLabel.Text" xml:space="preserve"><value>按作者名筛选:</value></data>
-  <data name="FilterByAuthorTextBox.Location" type="System.Drawing.Point, System.Drawing"><value>543, 74</value></data>
-  <data name="FilterByAuthorTextBox.Size" type="System.Drawing.Size, System.Drawing"><value>170, 26</value></data>
-  <data name="FilterByNameLabel.Text" xml:space="preserve"><value>按Mod名称筛选:</value></data>
-  <data name="FilterByNameTextBox.Location" type="System.Drawing.Point, System.Drawing"><value>160, 74</value></data>
-  <data name="FilterByNameTextBox.Size" type="System.Drawing.Size, System.Drawing"><value>170, 26</value></data>
-  <data name="FilterByDescriptionLabel.Text" xml:space="preserve"><value>按描述筛选:</value></data>
-  <data name="FilterByDescriptionTextBox.Location" type="System.Drawing.Point, System.Drawing"><value>912, 74</value></data>
-  <data name="FilterByDescriptionTextBox.Size" type="System.Drawing.Size, System.Drawing"><value>170, 26</value></data>
   <data name="ChangesetTabPage.Text" xml:space="preserve"><value>更改列表</value></data>
   <data name="WaitTabPage.Text" xml:space="preserve"><value>状态日志</value></data>
   <data name="ChooseProvidedModsTabPage.Text" xml:space="preserve"><value>选择Mod</value></data>
@@ -243,6 +199,4 @@
   <data name="openKSPDirectoryToolStripMenuItem1.Text" xml:space="preserve"><value>打开KSP目录</value></data>
   <data name="cKANSettingsToolStripMenuItem1.Text" xml:space="preserve"><value>CKAN设置</value></data>
   <data name="quitToolStripMenuItem.Text" xml:space="preserve"><value>退出</value></data>
-  <data name="labelsToolStripMenuItem.Text" xml:space="preserve"><value>标签</value></data>
-  <data name="editLabelsToolStripMenuItem.Text" xml:space="preserve"><value>编辑标签...</value></data>
 </root>
