+++ GUI/Main/MainImport.cs
@@ -39,7 +39,7 @@ namespace CKAN
                     ModuleInstaller.GetInstance(CurrentInstance, Manager.Cache, currentUser).ImportFiles(
                         GetFiles(dlg.FileNames),
                         currentUser,
-                        (CkanModule mod) => MarkModForInstall(mod.identifier, false),
+                        (CkanModule mod) => ManageMods.MarkModForInstall(mod.identifier, false),
                         RegistryManager.Instance(CurrentInstance).registry
                     );
                 }
