+++ Tests/GUI/ModList.cs
@@ -13,32 +13,20 @@ using ModuleInstaller = CKAN.ModuleInstaller;
 namespace Tests.GUI
 {
     [TestFixture]
-    public class MainModListTests
+    public class ModListTests
     {
         [Test]
         public void OnCreation_HasDefaultFilters()
         {
-            var item = new MainModList(delegate { });
+            var item = new ModList(delegate { });
             Assert.AreEqual(GUIModFilter.Compatible, item.ModFilter, "ModFilter");
-            Assert.AreEqual(String.Empty, item.ModNameFilter, "ModNameFilter");
         }
 
-        [Test]
-        public void OnModTextFilterChanges_CallsEventHandler()
-        {
-            var called_n = 0;
-            var item = new MainModList(delegate { called_n++; });
-            Assert.That(called_n == 1);
-            item.ModNameFilter = "randomString";
-            Assert.That(called_n == 2);
-            item.ModNameFilter = "randomString";
-            Assert.That(called_n == 2);
-        }
         [Test]
         public void OnModTypeFilterChanges_CallsEventHandler()
         {
             var called_n = 0;
-            var item = new MainModList(delegate { called_n++; });
+            var item = new ModList(delegate { called_n++; });
             Assert.That(called_n == 1);
             item.ModFilter = GUIModFilter.Installed;
             Assert.That(called_n == 2);
@@ -49,7 +37,7 @@ namespace Tests.GUI
         [Test]
         public void ComputeChangeSetFromModList_WithEmptyList_HasEmptyChangeSet()
         {
-            var item = new MainModList(delegate { });
+            var item = new ModList(delegate { });
             Assert.That(item.ComputeUserChangeSet(null), Is.Empty);
         }
 
@@ -68,7 +56,7 @@ namespace Tests.GUI
                 var ckan_mod = TestData.FireSpitterModule();
                 var registry = Registry.Empty();
                 registry.AddAvailable(ckan_mod);
-                var item = new MainModList(delegate { });
+                var item = new ModList(delegate { });
                 Assert.That(item.IsVisible(
                     new GUIMod(ckan_mod, registry, manager.CurrentInstance.VersionCriteria()),
                     manager.CurrentInstance.Name
@@ -81,7 +69,7 @@ namespace Tests.GUI
         [Test]
         public void CountModsByFilter_EmptyModList_ReturnsZeroForAllFilters()
         {
-            var item = new MainModList(delegate { });
+            var item = new ModList(delegate { });
             foreach (GUIModFilter filter in Enum.GetValues(typeof(GUIModFilter)))
             {
                 Assert.That(item.CountModsByFilter(filter), Is.EqualTo(0));
@@ -104,7 +92,7 @@ namespace Tests.GUI
                 var registry = Registry.Empty();
                 registry.AddAvailable(TestData.FireSpitterModule());
                 registry.AddAvailable(TestData.kOS_014_module());
-                var main_mod_list = new MainModList(null);
+                var main_mod_list = new ModList(null);
                 var mod_list = main_mod_list.ConstructModList(
                     new List<GUIMod>
                     {
